[FIX] sale_timesheet: correctly trigger compute
Revision 161429e9a7efe6cc9785012ccdc33a0bbacb643b introduced a forced
recompute of `billable_type` which unfortunately has repercutions
(compute triggers) upon every single task of a project, meaning that it
will not scale properly as the number of tasks grows and it is now
called everytime we want to access a project view form whch is a
terrible performance drop.

The underlying issue that it tried to address was that the field
billable_type was not correctly computed when deleting the linked
sale_order_id from the project. This happens because when deleting a
record, it is PSQL that sets relational m2o value (in this cas
sale_order_id to NULL and the ORM has no way to catch that information.

A proper fix consists to change the dependency of the computed field
from the m2o to one of its fields. This will properly mark
'billable_type' to be recomputed when unlinking the sale order.

Note that to avoid useless recomputation which would lead to the sale
result, we rely on a filed that should not change once the sale order is
created.

closes odoo/odoo#54817

X-original-commit: 591a5d7cbded0d564fca993d5fafad3a2f9d78e7
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>