[FIX] web_editor: restore custom snippet delete button
Clicking on the delete button of custom snippets was not working anymore
since [1]. Indeed, that button is placed inside the snippet thumbnail
which is the drag handle for the drag and drop feature. Since we now
start the drag as soon as the mouse is mousedown-ed, then the click on
the delete button cannot be intercepted.

This commit solves the problem by placing the delete button out of the
thumbnail element.

[1]: https://github.com/odoo/odoo/commit/a5bd196796641fc88705c54e3764a1d333ccd3f7

closes odoo/odoo#54770

X-original-commit: f50539121232a9e1c5e7066c800dd05d7268bfa3
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>