[FIX] website_sale: Fetch pricelist with proper rights
When a user with access to company A tried to make a purchase on the
website of company B, he would face a traceback upon calling `_get_pricelist_available`.
This happens because we eventually try to access to the current company
(B) while forcing environment of the user (in company A), thus
forfeiting the superuser privileges, which triggers
an access error.

closes odoo/odoo#54891

X-original-commit: 44b156bf3d569f76aa6149b35f3a4efac92e443d
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>