[FIX] website_sale: align abandoned orders count and records
The purpose of the button "Abandoned Carts" on the eCommerce dashboard is to
encourage user to send recovery emails.
Before this commit, the domain wasn't correct as we considered all the
abandoned cart, including the ones which an email was sent.

After this commit, the domain of the filter "Abandoned Carts" took into
consideration the abandoned cart and the carts for whom it is necessary
to send an email.

This commit closes #52799