[FIX] hr_timesheet: Timesheet in multi company
Allow user that have a employee in a company A
and that is allowed on company A and B to timesheet
on company B

Problem:
A user have access on company A and B but is link to
an employee that belong to company A. He want to timesheet on a
project from the company B. He gets an error
"Incompatible companies on records".

Since most users are linked to only one employee that belongs
to one company. It's impossible to have consistent companies.

Solution:

The employee shouldn't be checked on the timesheet

closes odoo/odoo#55074

X-original-commit: dc15e7d29b5bb17b3489a429ada3c123d65b45f4
Signed-off-by: Olivier Dony (odo) <odo@openerp.com>