[FIX] hr_recruitment: Include archived applicants on job's button count
Currently in the job's stat button and on kanban button the archived
applicants are not counted and not display while click on button.

So in this commit, count of the  archived applicant's in form views
and show the archived applicants when clicked on stat button.
also curently there is no indication for applicant is refused or not
so in this commit add badge on kanban and form for archieve(refused)
applicant.
we should not count the archived applicantion in the kanban
button.

closes odoo/odoo#55755

Taskid: 2308448
Closes: #55104
X-original-commit: b7ce6ab218674cd8fe6f23fb0b43a5445c5d565c
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>