[FIX] website: prevent page flickering when modal is opened
Commit [1] introduced a flickering when any modal was opened on the
frontend. Indeed, it introduced the possibility to go into "fullscreen"
mode, which makes the Odoo UI navbar and sidebar disappear. This is
animated: the padding of the `<body/>` element is using a CSS
transition... but when a Bootstrap modal opens, the padding-right is
changed to compensate the scrollbar removal. This change now being
animated, it creates a flickering.

A generic solution was made: only animate what is animated for the
duration of the fullscreen mode switch.

[1]: https://github.com/odoo/odoo/commit/d4e7a71a288f1cd86b6ec97903585e156239bd55

closes odoo/odoo#55210

X-original-commit: 14061af36617f46b4c49f0bc0e18ed4825ec03d5
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>