[IMP] stock_picking_batch: internal -> delivery
Change the demo data to create pickings/picking batchs about deliveries
instead of internal transfers.
Made this change because multi-location is disabled by default so it
doesn't make sense to work with internal transfer with this config.
Also, renames these data records reference to avoid migration issue as
the pickings/picking batches picking type was changed.

Adds a second picking in the batch picking BATCH/00001 because it
doesn't make sense to have a batch picking for only one picking.

X-original-commit: ff2d4d77019a137008a7af5b0742c9ababeb5df9