[FIX] sale_timesheet: Create invoice from SO with section
Problem:
When you try to create an invoice from a sale order that contains
section and post paid service product.
You set a date_invoice_timesheet in the wizard.

You got a traceback excpected singleton product.product()
This is cause by the filtering of so line with
sol.product_id._is_delivered_timesheet()
when the sale order line has no product for a section or a note

Solution:
Check there is a product on the line first

closes odoo/odoo#55353

X-original-commit: e02c38c0bfba278ff9ebe9059e4e4b4261d64a12
Signed-off-by: Olivier Dony (odo) <odo@openerp.com>
Signed-off-by: Thibault Francois <tfr-odoo@users.noreply.github.com>