[IMP] web_editor: remove check icon for default active buttons
Before this commit, the enabled state of we-button element was a
"pressed-like" background + a check icon. This check icon was then
explicitely removed for we-button extension such as checkboxes and
button groups' buttons. Now, the enabled state does not contains that
check icon by default and it is explicitely added for selects' buttons.

Part of https://github.com/odoo/odoo/pull/53017
task-2210790